made registerscreen work
