+++ src/screens/RegisterScreen/RegisterScreen.js
@@ -1,9 +1,54 @@
-import React from "react";
+import { React, useState } from "react";
 import Styles from "./Styles";
-import { View, Text, SafeAreaView } from "react-native";
-import { Button } from "react-native-web";
+import { View, Text, ScrollView, TextInput, Pressable, Image, SafeAreaView } from "react-native";
+import { useValidation } from "react-native-form-validator";
 
-const RegisterScreen = ({ navigation, route }) => {
+const RegisterScreen = ({ navigation }) => {
+    const [name, setName] = useState("");
+    const [email, setEmail] = useState("");
+    const [date, setDate] = useState("");
+    const [newPassword, setNewPassword] = useState("");
+    const [confirmPassword, setConfirmPassword] = useState("");
+
+    const { validate, isFieldInError, getErrorsInField, isFormValid } =
+    useValidation({
+      state: { name, email, date, newPassword, confirmPassword },
+    });
+
+    /**
+     * 
+     */
+    const onPressButton = () => {
+        validate({
+          name: { required: true, maxLength: 50 },
+          email: { email: true, required: true, maxLength: 50 },
+          date: { date: "YYYY-MM-DD", required: true },
+          newPassword: { required: true, isPassword: true },
+          confirmPassword: { equalPassword: newPassword, required: true },
+        }),
+
+        checkValidation();
+          // () => {
+          //   if (isFormValid() == true) {
+          //       sendDataToAPI(name, email, date, newPassword, confirmPassword);
+          //   }
+          // };
+    };
+
+    const checkValidation = () => {
+        if (isFormValid() == true) {
+            sendDataToAPI(name, email, date, newPassword, confirmPassword);
+        }
+    };
+
+    /**
+     * 
+     * @param {*} name
+     * @param {*} email
+     * @param {*} dateOfBirth
+     * @param {*} password
+     * @param {*} confirmPassword
+     */
 	const sendDataToAPI = (name, email, dateOfBirth, password, confirmPassword) => {
         try {
             fetch("http://localhost/pma/PmaAPI/handlers/registration/registrationHandler.php", {
@@ -32,74 +77,127 @@ const RegisterScreen = ({ navigation, route }) => {
 	};
 
     /**
-     * Function to catch feedback of request to API
-     * @param response JSON object of response
+     * Catch feedback of request from API
+     * @param {*} response 
      */
-    const catchFeedback = (response) => {
-        
+     const catchFeedback = (response) => {
         // console.log(response);
-        // console.log(response.length);
-
-        for (let index = 0; index < response.length; index++) {
-            switch (response[index]) {
-                case 'name_incorrect':
-                    //do something
-                    console.log('name_incorrect');
-                    break;
-
-                case 'email_incorrect':
-                    //do something
-                    console.log('email_incorrect');
-                    break;
-
-                case 'dateOfBirth_incorrect':
-                    //do something
-                    console.log('dateOfBirth_incorrect');
-                    break;
-
-                case 'password_incorrect':
-                    //do something
-                    console.log('password_incorrect');
-                    break;
-
-                case 'confirmPassword_incorrect':
-                    //do something
-                    console.log('confirmPassword_incorrect');
-                    break;
-
-                case 'samePassword_incorrect':
-                    //do something
-                    console.log('samePassword_incorrect');
-                    break;
-
-                case 'email_in_use':
-                    //do something
-                    console.log('email_in_use');
-                    break;
-                
-                default:
-                    //Succes
-                    console.log(response[0].id);
-                    console.log(response[0].name);
-                    console.log(response[0].email);
-                    console.log(response[0].dateOfBirth);
-                    break;
-            }
+    
+        switch (response[0]) {
+          case "name_incorrect":
+            alert("De naam is verkeerd.");
+            break;
+          case "email_incorrect":
+            alert("Het emailadres is verkeerd.");
+            break;
+          case "dateOfBirth_incorrect":
+            alert("De geboortedatum is verkeerd.");
+            break;
+          case "password_incorrect":
+            alert("Het eerste wachtwoord is verkeerd.");
+            break;
+          case "confirmPassword_incorrect":
+            alert("Het tweede wachtwoord is verkeerd.");
+            break;
+          case "samePassword_incorrect":
+            alert("De wachtwoorden zijn niet hetzelfde.");
+            break;
+          case "email_in_use":
+            alert("Dit emailadres is al in gebruik.");
+            break;
+          default:
+            alert("Je account is aangemaakt, log nu in met je gegevens.");
+            navigation.navigate('LoginScreen');
         }
-    };
+      };
 
-	return (
-		<SafeAreaView style={Styles.SafeAreaView}>
-			<View style={Styles.head}>
-				<Text>logo</Text>
-			</View>
-
-			<View style={Styles.login}>
-				<Text>Hier alles voor registratie</Text>
-                <Button title="druk hier" onPress={() => sendDataToAPI("Stefan", "stefan@email.com", "2001-12-26", "!Welkom10", "!Welkom10")} />
-			</View>
-		</SafeAreaView>
-	);
-}
+    return (
+        <SafeAreaView style={Styles.container}>
+          <View>
+            <Image
+              style={Styles.logo}
+              source={require("../../assets/images/logo.png")}
+            />
+          </View>
+    
+          <Text style={Styles.titel}>REGISTREREN</Text>
+    
+          <View style={Styles.div}>
+            <TextInput
+              style={Styles.input}
+              onChangeText={setName}
+              value={name}
+              placeholder="Naam"
+            />
+            {isFieldInError("name") &&
+              getErrorsInField("name").map((errorMessage) => (
+                <Text style={Styles.text}>{errorMessage}</Text>
+              ))}
+          </View>
+    
+          <View style={Styles.div}>
+            <TextInput
+              style={Styles.input}
+              onChangeText={setEmail}
+              value={email}
+              placeholder="Email"
+            />
+            {isFieldInError("email") &&
+              getErrorsInField("email").map((errorMessage) => (
+                <Text style={Styles.text}>{errorMessage}</Text>
+              ))}
+          </View>
+    
+          <View style={Styles.div}>
+            <TextInput
+              style={Styles.input}
+              onChangeText={setDate}
+              value={date}
+              placeholder="Datum"
+            />
+            {isFieldInError("date") &&
+              getErrorsInField("date").map((errorMessage) => (
+                <Text style={Styles.text}>{errorMessage}</Text>
+              ))}
+          </View>
+    
+          <View style={Styles.div}>
+            <TextInput
+              style={Styles.input}
+              onChangeText={setNewPassword}
+              value={newPassword}
+              secureTextEntry={true}
+              placeholder="Wachtwoord"
+            />
+            {isFieldInError("newPassword") &&
+              getErrorsInField("newPassword").map((errorMessage) => (
+                <Text style={Styles.text}>{errorMessage}</Text>
+              ))}
+          </View>
+    
+          <View style={Styles.div}>
+            <TextInput
+              style={Styles.input}
+              onChangeText={setConfirmPassword}
+              value={confirmPassword}
+              secureTextEntry={true}
+              placeholder="Confirm wachtwoord"
+            />
+            {isFieldInError("confirmPassword") &&
+              getErrorsInField("confirmPassword").map((errorMessage) => (
+                <Text style={Styles.text}>{errorMessage}</Text>
+              ))}
+          </View>
+    
+          <Pressable onPress={onPressButton}>
+            <Text style={Styles.button}>REGISTREREN</Text>
+          </Pressable>
+    
+          <Pressable onPress={() => navigation.navigate("LoginScreen")}>
+            <Text style={Styles.inloggen}>INLOGGEN</Text>
+          </Pressable>
+        </SafeAreaView>
+      );
+};
 
 export default RegisterScreen;
